pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mail-bridge"
version = "0.1.0"
description = "Gmail push bridge for importance classification and QQ notifications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0",
  "google-api-python-client>=2.170.0",
  "google-auth>=2.39.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.2",
  "google-cloud-pubsub>=2.29.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.8.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["mail_bridge*"]
